requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
coverage==7.2.7
Django==4.2.2
django-cors-headers==4.1.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
importlib-metadata==6.7.0
Markdown==3.4.3
phonenumbers==8.13.15
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.7.0
urllib3==2.0.3
yarg==0.1.9
zipp==3.15.0
